tutorial refactored - some fixes

+++ app/src/de/ur/mi/android/ting/model/dummy/DummyConfig.java
@@ -1,7 +1,7 @@
 package de.ur.mi.android.ting.model.dummy;
 
 public class DummyConfig {
-	public static final long DUMMY_SIMULATED_NETWORK_DELAY_inmilliseconds = 1000;
+	public static final int DUMMY_SIMULATED_NETWORK_DELAY_inmilliseconds = 1000;
 	public static final boolean IS_USER_LOGGED_DEFAULT = false;
 	public static final String MEDIUM_LONG_TEXT = "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo.";
 }
